Moves constants to const.js
